Adopt InitializeRequestHandler
